requirements.txt
Django==3.2.4
requests==2.25.1
python-decouple==3.4
